requirements_flower.txt
tensorflow
numpy
opencv-python
matplotlib
scikit-learn
pillow
streamlit
